Attach sec groups to the VIP port on bastion
Heat doesn't allow to attach security groups to an instance when it also
has a port attached, and in order to get to the client those security
groups need to be applied on the second interface of the bastion host.

This change creates a port for the VIP network and attaches sec groups
to it.